pyproject.toml
[project]
name = "pog-pfn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "torch>=2.10.0",
    "tqdm>=4.67.2",
]
